move electro slash commands to electroslash cog
