+++ thatkitebot/cogs/electronics.py
@@ -26,7 +26,7 @@ import discord.commands as scmd
 from discord.ext import commands
 import si_prefix
 from random import randint
-from random import uniform
+
 from thatkitebot.backend import util
 
 
@@ -518,69 +518,6 @@ class ElectroCog(commands.Cog, name="Electronics commands"):
         except TooFewArgsError:
             await util.errormsg(ctx, "Not enough arguments to compute anything.")
             return
-
-    @scmd.slash_command(guild_ids=[759419755253465188], name="rc")
-    async def _rc(
-            self,
-            ctx: discord.ApplicationContext,
-            c1: scmd.Option(str, "Value for C1:", required=False, default=None),
-            r1: scmd.Option(str, "Value for R1", required=False, default=None),
-            fcut: scmd.Option(str, "cutoff frequency:", required=False, default=None),
-            draw_plot: scmd.Option(bool, "Display plot", required=False, default=False)
-    ):
-        """
-        Calculate different aspects of an RC filter.
-        Run the command for more details.
-        """
-        if not c1 and not r1 and not fcut:
-            random_rc = {
-                "fcut": str(uniform(0.1, 10 ** 5)),
-                "r1": str(uniform(100, 100000))
-            }
-            args_parsed = parse_input(random_rc)
-            rc = RCFilter(d=args_parsed, plot=False)
-            await ctx.respond(embed=rc.gen_embed())
-            return
-
-        args_parsed = dict(
-            c1=slash_preprocessor(c1),
-            r1=slash_preprocessor(r1),
-            fcut=slash_preprocessor(fcut)
-        )
-        try:
-            if draw_plot:
-                rc = RCFilter(d=args_parsed, plot=True)
-                await ctx.respond(embed=rc.gen_embed(), file=rc.gen_file())
-            else:
-                rc = RCFilter(d=args_parsed, plot=False)
-                await ctx.respond(embed=rc.gen_embed())
-
-        except TooFewArgsError:
-            a = await util.errormsg(ctx, "Not enough arguments to compute anything.", embed_only=True)
-            await ctx.respond(embed=a)
-            return
-
-    @scmd.slash_command(guild_ids=[759419755253465188], name="divider")
-    async def _divider(
-            self,
-            ctx: discord.ApplicationContext,
-            r1: scmd.Option(str, "Value for R1", required=False, default=None),
-            r2: scmd.Option(str, "Value for R2", required=False, default=None),
-            vin: scmd.Option(str, "Input voltage", required=False, default=None),
-            vout: scmd.Option(str, "Output Voltage", required=False, default=False)
-    ):
-        """
-        Calculate values of an unloaded voltage divider. Run the command for more details.
-        Thank you dimin for the idea and the "art"
-        """
-        args_parsed = dict(
-            r1=slash_preprocessor(r1),
-            r2=slash_preprocessor(r2),
-            vin=slash_preprocessor(vin),
-            vout=slash_preprocessor(vout)
-        )
-        div = VoltageDivider(d=args_parsed)
-        await ctx.respond(embed=div.gen_embed())
 
 
 def setup(bot):
